rewrite tests
